requirements.txt
# You must use python 3.13.3, to make this work right.
Flask==2.3.2
pandas==2.0.1
cassandra-driver==3.25.0
joblib==1.2.0
scikit-learn==1.2.2
nltk==3.7
kafka-python==2.0.2
tabulate==0.9.0